factor out provisioner call so it can be used in another freedom module
